Update the file name reprocessor

+++ lyricsFileNameReprocessor.ts
@@ -5,6 +5,9 @@ import * as process from 'process';
 import { TXT_EXTENSION } from './constants';
 import { normalizeFileName } from './src';
 import recursive from 'recursive-readdir';
+import { isEqual } from 'lodash';
+import { logFileWithLinkInConsole } from './utils';
+import chalk from 'chalk';
 
 dotenv.config();
 
@@ -17,23 +20,36 @@ const reprocessFileNames = async (dir: string) => {
   console.log(`"Reprocessing file names from ${dir} directory.."`);
 
   (await recursive(dir))
-    .filter((filePath) => path.extname(filePath) === TXT_EXTENSION)
+    .filter((filePath) => path.extname(filePath) === '')
     .forEach((filePath) => {
       const existingContent = fs.readFileSync(filePath).toString();
-
       const fileName = path.basename(filePath);
       const newFileName = normalizeFileName(fileName);
-      console.log(
-        `Processing "${fileName}": ${
-          fileName !== newFileName ? newFileName : 'No change.'
-        }.`,
-      );
+      const hasNoChange = isEqual(fileName, newFileName);
+
+      console.group(chalk.cyan(`Processing "${fileName}".`));
+      logFileWithLinkInConsole(filePath);
+
+      if (hasNoChange) {
+        console.log(chalk.yellow('Skipped the file.'));
+        console.log();
+        console.groupEnd();
+
+        return;
+      }
 
       fs.unlinkSync(filePath);
-      fs.writeFileSync(
-        path.join(__dirname, path.dirname(filePath), newFileName),
-        existingContent,
+
+      const nextPathName = path.join(
+        __dirname,
+        path.dirname(filePath),
+        newFileName,
       );
+      fs.writeFileSync(nextPathName, existingContent);
+
+      console.log(chalk.green(`Renamed to ${newFileName}.`));
+      console.log();
+      console.groupEnd();
     });
 };
